gpsd_client: try to fix /extended_fix handling
The previous version seems broken, with missing curly-braces breaking
the population of the values.

Here we try to fix this by just checking if we have a fix mode that is
2D-fix or better instead of checking the fix status from GPSD which can
be 0 (no fix or unknown fix status) even though we actually have a fix.
If this is the case, we populate the extended_fix message.